// src/types.ts
/* Copyright (c) 2021, VRAI Labs and/or its affiliates. All rights reserved.
 *
 * This software is licensed under the Apache License, Version 2.0 (the
 * "License") as published by the Apache Software Foundation.
 *
 * You may not use this file except in compliance with the License. You may
 * obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations
 * under the License.
 */

import RecipeModule from './recipeModule'
import NormalisedURLDomain from './normalisedURLDomain'
import NormalisedURLPath from './normalisedURLPath'
import { TypeFramework } from './framework/types'

export interface AppInfo {
  appName: string
  websiteDomain: string
  websiteBasePath?: string
  apiDomain: string
  apiBasePath?: string
  apiGatewayPath?: string
}

export interface NormalisedAppinfo {
  appName: string
  websiteDomain: NormalisedURLDomain
  apiDomain: NormalisedURLDomain
  topLevelAPIDomain: string
  topLevelWebsiteDomain: string
  apiBasePath: NormalisedURLPath
  apiGatewayPath: NormalisedURLPath
  websiteBasePath: NormalisedURLPath
}

export interface SuperTokensInfo {
  connectionURI: string
  apiKey?: string
}

export interface TypeInput {
  supertokens?: SuperTokensInfo
  framework?: TypeFramework
  appInfo: AppInfo
  recipeList: RecipeListFunction[]
  telemetry?: boolean
  isInServerlessEnv?: boolean
}

export type RecipeListFunction = (appInfo: NormalisedAppinfo, isInServerlessEnv: boolean) => RecipeModule

export interface APIHandled {
  pathWithoutApiBasePath: NormalisedURLPath
  method: HTTPMethod
  id: string
  disabled: boolean
}

export type HTTPMethod = 'post' | 'get' | 'delete' | 'put' | 'options' | 'trace'

export type JSONPrimitive = string | number | boolean | null
export type JSONArray = Array<JSONValue>
export type JSONValue = JSONPrimitive | JSONObject | JSONArray | undefined
export interface JSONObject {
  [ind: string]: JSONValue
}
export interface GeneralErrorResponse {
  status: 'GENERAL_ERROR'
  message: string
}